Handle HTTPS connections to GSD
This does introduce a little ugly method to accept the certificate for
clients, requires them to click a link to go to the GSD server to
accept the certificate. No way around this without a signed certificate
(which can’t be done for local IPs, or for IPs in general I believe).
Will require rewriting some code to connect to GSD using a domain in
the future to support signed certificates.