Fixed URL encoding issue for file upload (#788)
This commit fixes an issue where uploaded filenames have characters
requiring URL encoding.

The proxy failed when the filename had characters (such as a space) that
need to be escaped.  The filename portion of the request was not
properly encoded and that caused a server error in pufferd (as it ended
up being an invalid request that could not be parsed).