Arch Linux Support + Refactoring (#800)
* auto-deploy added support for Arch

* Replaced stray 'print' commands with 'echo'

* pufferpanel mode 754

* Added ROW_FORMAT to prevent ERR1071 key-length

ROW_FORMAT is presumed to be Dynamic for certain versions of InnoDB;
however, this is not always the case. If ROW_FORMAT is not correctly set
then innodb_large_prefix will not work regardless of being set to 1.

* php/php-fpm.d/www.conf support for NGinx PhpSocket

* Utility function for Php www conf location

* Refactored php-fpm path usage

Replaced all indent hadouken with single greps, using php-fpm path
obtained from getPhpConfLocation utility function.

* PufferPanel NGinx Restart systemctl support

Added a switch to check for systemctl or service, to use the appropriate
restart command

* PufferPanel Apache Restart Systemctl Support

Added a switch to check for systemctl or service, to use the appropriate
restart command.
Additionally, changed the Apache service name to correct httpd for
systemd service file.